+++ sdk/schemaregistry/azure-schemaregistry/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.2.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.2.0 (2022-10-10)
 
 This version and all future versions will require Python 3.7+, Python 3.6 is no longer supported.